Consolidate channel viewsets. Move to viewsets folder.
