+++ contentcuration/contentcuration/urls.py
@@ -20,18 +20,10 @@ from django.conf.urls import url
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
 from django.core.urlresolvers import reverse_lazy
-from django.db.models import Count
-from django.db.models import IntegerField
-from django.db.models import Max
-from django.db.models import OuterRef
-from django.db.models import Prefetch
 from django.db.models import Q
-from django.db.models import Subquery
-from django_filters.rest_framework import BooleanFilter
 from django_filters.rest_framework import CharFilter
 from django_filters.rest_framework import DjangoFilterBackend
 from django_filters.rest_framework import FilterSet
-from le_utils.constants import content_kinds
 from rest_framework import permissions
 from rest_framework import routers
 from rest_framework import viewsets
@@ -51,6 +43,8 @@ import contentcuration.views.public as public_views
 import contentcuration.views.settings as settings_views
 import contentcuration.views.users as registration_views
 import contentcuration.views.zip as zip_views
+
+from contentcuration.viewsets.channel import ChannelViewSet
 from contentcuration.celery import app
 from contentcuration.forms import ForgotPasswordForm
 from contentcuration.forms import LoginForm
@@ -64,11 +58,9 @@ from contentcuration.models import ContentTag
 from contentcuration.models import File
 from contentcuration.models import FileFormat
 from contentcuration.models import FormatPreset
-from contentcuration.models import get_channel_thumbnail
 from contentcuration.models import Invitation
 from contentcuration.models import Language
 from contentcuration.models import License
-from contentcuration.models import SecretToken
 from contentcuration.models import Task
 from contentcuration.models import User
 
@@ -90,155 +82,6 @@ class LanguageViewSet(viewsets.ModelViewSet):
     queryset = Language.objects.all()
 
     serializer_class = serializers.LanguageSerializer
-
-
-class ChannelViewSet(viewsets.ModelViewSet):
-    queryset = Channel.objects.all()
-    serializer_class = serializers.ChannelSerializer
-
-    def get_queryset(self):
-        if self.request.user.is_admin:
-            return Channel.objects.all()
-        return Channel.objects.filter(Q(editors=self.request.user) | Q(viewers=self.request.user) | Q(public=True)).distinct()
-
-
-class ChannelFilter(FilterSet):
-    edit = BooleanFilter(method="filter_edit")
-    view = BooleanFilter(method="filter_view")
-    bookmark = BooleanFilter(method="filter_bookmark")
-    ids = CharFilter(method="filter_ids")
-
-    class Meta:
-        model = Channel
-        fields = ("edit", "view", "public", "bookmark", "ids")
-
-    def filter_edit(self, queryset, name, value):
-        return queryset.filter(editors=self.request.user)
-
-    def filter_view(self, queryset, name, value):
-        return queryset.filter(viewers=self.request.user)
-
-    def filter_bookmark(self, queryset, name, value):
-        return queryset.filter(bookmarked_by=self.request.user)
-
-    def filter_ids(self, queryset, name, value):
-        try:
-            # Limit SQL params to 50 - shouldn't be fetching this many
-            # ids at once
-            return queryset.filter(pk__in=value.split(",")[:50])
-        except ValueError:
-            # Catch in case of a poorly formed UUID
-            return queryset.none()
-
-
-class SQCount(Subquery):
-    # Include ALIAS at the end to support Postgres
-    template = "(SELECT COUNT(%(field)s) FROM (%(subquery)s) AS %(field)s__sum)"
-    output_field = IntegerField()
-
-
-class ChannelSlimViewSet(viewsets.GenericViewSet):
-    filter_backends = (DjangoFilterBackend,)
-    filter_class = ChannelFilter
-
-    def _map_fields(self, channel):
-        channel["thumbnail_url"] = get_channel_thumbnail(channel)
-        channel["published"] = channel["main_tree__published"]
-        channel.pop("main_tree__published")
-        channel["created"] = channel["main_tree__created"]
-        channel.pop("main_tree__created")
-        for key in ["edit", "view", "bookmark"]:
-            channel[key] = bool(channel[key])
-        return channel
-
-    def _annotate_queryset(self, queryset):
-        prefetch_secret_token = Prefetch(
-            "secret_tokens", queryset=SecretToken.objects.filter(is_primary=True)
-        )
-        prefetch_editors = Prefetch(
-            "editors", queryset=User.objects.filter(id=self.request.user.id)
-        )
-        prefetch_viewers = Prefetch(
-            "viewers", queryset=User.objects.filter(id=self.request.user.id)
-        )
-        prefetch_bookmarked = Prefetch(
-            "bookmarked_by", queryset=User.objects.filter(id=self.request.user.id)
-        )
-
-        queryset = queryset.select_related("language", "main_tree").prefetch_related(
-            prefetch_secret_token,
-            prefetch_editors,
-            prefetch_viewers,
-            prefetch_bookmarked,
-        )
-
-        queryset = queryset.annotate(primary_token=Max("secret_tokens__token"))
-        channel_main_tree_nodes = ContentNode.objects.filter(
-            tree_id=OuterRef("main_tree__tree_id")
-        )
-        # Add the last modified node modified value as the channel last modified
-        queryset = queryset.annotate(
-            modified=Subquery(
-                channel_main_tree_nodes.order_by("-modified").values("modified")[:1]
-            )
-        )
-        # Add the unique count of distinct non-topic node content_ids
-        non_topic_content_ids = (
-            channel_main_tree_nodes.exclude(kind_id=content_kinds.TOPIC)
-            .order_by("content_id")
-            .distinct("content_id")
-            .values_list("content_id", flat=True)
-        )
-        queryset = queryset.annotate(
-            count=SQCount(non_topic_content_ids, field="content_id")
-        )
-        # Annotate edit, view, and bookmark onto the channels
-        queryset = queryset.annotate(
-            edit=Count("editors"),
-            view=Count("viewers"),
-            bookmark=Count("bookmarked_by"),
-        )
-        return queryset
-
-    def _serialize_queryset(self, queryset):
-        queryset = self._annotate_queryset(queryset)
-        return queryset.values(
-            "id",
-            "name",
-            "description",
-            "main_tree__published",
-            "thumbnail",
-            "thumbnail_encoding",
-            "language",
-            "primary_token",
-            "modified",
-            "count",
-            "view",
-            "edit",
-            "bookmark",
-            "public",
-            "version",
-            "main_tree__created",
-            "last_published",
-            "ricecooker_version",
-        )
-
-    def get_queryset(self):
-        queryset = Channel.objects.filter(
-            Q(editors=self.request.user) | Q(viewers=self.request.user) | Q(public=True)
-        ).distinct()
-        return queryset
-
-    def list(self, request, *args, **kwargs):
-        queryset = self.filter_queryset(self.get_queryset())
-
-        page = self.paginate_queryset(queryset)
-        if page is not None:
-            data = map(self._map_fields, self._serialize_queryset(page) or [])
-            return self.get_paginated_response(data)
-
-        data = map(self._map_fields, self._serialize_queryset(queryset) or [])
-        return Response(data)
 
 
 class ChannelSetViewSet(viewsets.ModelViewSet):
@@ -436,7 +279,6 @@ router = routers.DefaultRouter(trailing_slash=False)
 router.register(r'license', LicenseViewSet)
 router.register(r'language', LanguageViewSet)
 router.register(r'channel', ChannelViewSet)
-router.register(r'channelslim', ChannelSlimViewSet, base_name="channelslim")
 router.register(r'channelset', ChannelSetViewSet)
 router.register(r'fileformat', FileFormatViewSet)
 router.register(r'preset', FormatPresetViewSet)
